+++ R/backend-snowflake.R
@@ -36,6 +36,24 @@ sql_translation.Snowflake <- function(con) {
         }
         sql_call2("CONCAT_WS", sep, ...)
       },
+      str_like = function(string, pattern, ignore_case = deprecated()) {
+
+        if (lifecycle::is_present(ignore_case)) {
+          lifecycle::deprecate_warn(
+            when = "1.6.0",
+            what = "str_like(ignore_case)",
+            details = c(
+              "`str_like()` is always case sensitive.",
+              "Use `str_ilike()` for case insensitive string matching."
+            )
+          )
+          check_bool(ignore_case)
+          if (ignore_case) {
+            return(sql_expr(!!string %ILIKE% !!pattern))
+          }
+        }
+        sql_expr(!!string %LIKE% !!pattern)
+      },
       str_ilike = function(string, pattern) {
         sql_expr(!!string %ILIKE% !!pattern)
       },